Update multiple-error-types section for new flow
`option_result.md` is a new page for showing the naiive approach while
also included the trick for swapping `Option<Result>` to
`Result<Optipn>`.

`boxing_errors.md` was moved earlier in the material to quickly get the
reader to something they can use to get decent results in their
application with minimal fuss. This was also part of an effort to remove
the use of `Result<_, String>` to avoid the issue of someone jumping
into a section they need, seeing it, and using it when there are much
better alternatives.

`define_error_type.md`s focus was narrowed to just defining errors,
leaving wrapping errors to the new `wrap_error.md` page.  In addition,
some basic traits were added to try to make the example code sufficient
for someone jumping into that page just to see how to define an error.

Moving `wrap_error.md` to the end also allowed some of the earlier
examples to include less boilerplate, minimizing distraction from the
point that is trying to be made.

Links and examples have been tested.

Fixes #901, #902