freeze.md: Incorrect example
The freeze example is meant to show that a value cannot be immutably borrowed
while the original value is modified within the same scope.

However, as the immutably borrowed value is not used after the mutation, the
borrow checker allows the example to be compiled successfully, contrary to what
we intend to demonstrate.

Solution: Use the immutably borrowed value after original value is modified.