Fight the confusion over 'static
Many people learning Rust are confused by the fact that owned values are `'static`. I feel the description here as well as in _the book_ are responsible for that. 

I hope that this change makes things a lot clearer. Note that I have intersected the explanation before the existing example because if It's underneath, it will 
be offscreen without scrolling. I kind of feel this is to important to be hidden offscreen, even if I tried to reformulate the original text to be more accurate.

I would propose we put this explanation in _the book_ as well, in the section about 'static lifetimes.

For an example of what can happen if people are confused about lifetime bounds (and use unsafe): https://github.com/libra/libra/pull/1949