Moved "See also" link to the correct page.
The "8.5.1.3. pointers/ref" code example contains multiple usages of the
ref pattern, however no link to it is included in the "See also" section
at the end of the page (the section is missing). In the next page,
"8.5.1.4. structs", the code example contains no usage of the ref
pattern, however its "See also" section contains a link to it. This
commit moves the ref pattern link to the previous page: "8.5.1.3.
pointers/ref" to remedy the problem.