Clarify distinction between for iter and into_iter
The existing code examples for `for iter` and `for into_iter` don't make the distinction very clear to the user. 

Although this is explained at the end, allowing the user experiment with the match types and showing how the collection gets consumed in `into_iter` but not in `iter` would help them understand better.