Enums: Linked-List Needs Re-Wording
https://doc.rust-lang.org/rust-by-example/custom_types/enum/testcase_linked_list.html

I had a colleague ask me a question by being confused with the first sentence in this chapter:

> A common use for enums is to create a linked-list:

IMO, this should be the other way around. A common way to implement Linked-Lists
is to use enums. But I find it not aligned with my experience that a common use
of enums are to create linked lists.