store Person.age on the heap
I think that `ref` is not required in this example if Person.age is stored on the stack, as the definition of `age` copies the data instead of moving it. I have thus changed the code to store it on the heap, to illustrate the usefulness of `ref` in this context, and added a short explanation. 

An alternative possibility could be to use `name` instead of `age` for the illustration; not sure which is clearer. Please feel free to edit or request edits to this PR if you see a clearer way!