Use rust 2018 module file hierarchy
With 2018 edition we no longer need mod.rs files. The book follows the
new convention in its dedicated chapter and the reference encourages its
usage over the mod.rs files. In order to be consistent with the book, we
now have `my.rs` and `my` folder in the same directory instead of
`my/mod.rs`.