Update print.md (#1597)
On the website, there are two options listed for hexadecimal. I was confused what the difference between the capital x and X was but then I ran it through and realized the latter capitalized the letters. I've also done the courtesy of listing out their subsequent results.

I also added some additional information to the "width" section and formatted it for easier readability.