Misleading textual statement in HOF
The example checks if the calculated square is less than 1000, and if it
is odd, then adds it to the sum - not if the not squared number is odd.
The amount of calculations is much lower than if it was checking if the
not squared number is less than 1000 and then square it and add it to
the sum.