Update while_let.md: address inconsistent use of fn main between 2 compared code examples
The original document compares 2 code blocks to illustrate the benefits of while let. The first example **is not** placed in a main function while the second example **is** placed in a main function.

The problem: This inconsistency is significant because, among other points of comparison, they are being compared in terms of their level of indentation.
Therefore...
Placing one example in a main function and not the other presents a visually unequal basis of comparison.

The solution I propose: is to just remove the main function in the second block since removing it reduces visual complexity.

I think adding in a main function to the first block and keeping the main function in the second block is also a good option. IMO consistency between the examples is what's most important.