fix(arg): Remove reference to Rust Cookbook in arg parsing
This PR removes the reference to the Rust Cookbook in the argument
parsing section. The example linked is outdated and no longer works.
Additionally, the repo hasn't seen any activity in a few years and
rust-lang-nursery has been deprecated anyway
(https://internals.rust-lang.org/t/rust-lang-nursery-deprecation/11205).

I chose to replace it with the `rust-cli` book link that also uses a
`clap` example.