Update the rustdoc example to match best practice
This change updates the meta/doc example to match the best practices
outlined in RFC 1574 (already referenced by the doc). Specifically, it
removes the "Arguments" section from the `Person` doc example.

I've also added a mention of `cargo doc --open` as it's pretty useful
and people may not think to see if it exists.